ready

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: mbotes <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/17 09:26:15 by mbotes            #+#    #+#             */
-/*   Updated: 2019/05/21 15:48:41 by mbotes           ###   ########.fr       */
+/*   Updated: 2019/08/13 14:24:58 by mbotes           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
-	int	sign;
-	int	res;
-	int	loop;
+	int			sign;
+	long long	res;
+	int			loop;
 
 	sign = 1;
 	res = 0;
